export helpButton to component
